requirements.txt
# Data
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0

# ML
scipy>=1.11.0
scikit-learn>=1.3.0
xgboost>=2.0.0
optuna>=3.4.0

# Technical Analysis
ta>=0.11.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0

# Utils
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Dev
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0
ruff>=0.1.0
